Require testfiles to provide valid input.
The reference implementation of bitarray_rotate has an assertion check
assert(bit_offset + bit_length <= bitarray->bit_sz). Its possible to write
a testfile which fails this assertion which can allow you to write tests
that have ambiguous results when run with assertions disabled. Before
running a test in a testfile, we now check whether this assertion holds.
If it doesn't, then we cause that test to fail and print an informative
error message.